adding docker and postgres support

+++ src/Database/CurrentRaidTier.cs
@@ -6,8 +6,8 @@ namespace NinjaBotCore.Database
     public class CurrentRaidTier
     {
         [Key]
-        public int Id { get; set; }
-        public int WclZoneId { get; set; }
+        public long Id { get; set; }
+        public long WclZoneId { get; set; }
         public string RaidName { get; set; }
         public int? Partition { get; set; }
     }
